+++ src/features/search/index.ts
@@ -1,2 +1,2 @@
 export { SearchResultsList } from "@/features/search/ui/SearchResultList";
-export { SearchUniversity } from "@/features/search/ui/searchUniv";
+export { SearchUniversity } from "@/features/search/ui/SearchUniversity";
